refactor: Remove RecyclerView and related logic from HomeFragment
