+++ app/build.gradle.kts
@@ -52,7 +52,6 @@ dependencies {
     implementation(libs.androidx.navigation.fragment.ktx)
     implementation(libs.androidx.navigation.ui.ktx)
     implementation(libs.androidx.core.splashscreen)
-    implementation(libs.kotlinx.datetime)
 
 
 }
