+++ app/src/main/java/com/samrraa/qurioapp/view/lastgames/model/History.kt
@@ -1,11 +1,9 @@
-package com.samrraa.qurioapp.view.home.model
-
-import kotlinx.datetime.*
+package com.samrraa.qurioapp.view.lastgames.model
 
 data class History(
     val date: String,
     val coin: Int,
     val star: String,
     val subject: String,
-    val time: LocalDateTime
+    val time: String
 )
